Update Account.java
Added two methods to search through an account's list of subscribers or list of subscriptions. The String parameter entered into each method should be the user name of the account you want to search for. If the String matches the user name of any of the accounts in the Subscribers or SubscribedTo list, the method returns true. Otherwise, it returns false.